Moved texture dealing methods to game area to help with refactoring
