+++ source/core/src/main/com/csse3200/game/areas/GameArea.java
@@ -1,15 +1,19 @@
 package com.csse3200.game.areas;
 
+import com.badlogic.gdx.audio.Music;
 import com.badlogic.gdx.math.GridPoint2;
 import com.badlogic.gdx.math.Vector2;
 import com.csse3200.game.areas.terrain.TerrainComponent;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.entities.factories.LoadedFactory;
+import com.csse3200.game.files.UserSettings;
+import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceLocator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
@@ -22,8 +26,8 @@ import java.util.concurrent.CopyOnWriteArrayList;
 public abstract class GameArea extends LoadedFactory {
     private static final Logger log = LoggerFactory.getLogger(GameArea.class);
     protected TerrainComponent terrain;
-
     protected List<Entity> areaEntities;
+    private static final String BACKGROUND_MUSIC = "sounds/BGM_03_mp3.mp3";
 
     protected GameArea() {
         super();
@@ -102,4 +106,86 @@ public abstract class GameArea extends LoadedFactory {
         CopyOnWriteArrayList<Entity> newAreaEntities = new CopyOnWriteArrayList<>(areaEntities);
         return newAreaEntities;
     }
+
+    //stuff moved from maingamearea??
+
+    protected void playBackgroundMusic() {
+        ResourceService resourceService = ServiceLocator.getResourceService();
+        if (!resourceService.containsAsset(BACKGROUND_MUSIC, Music.class)) {
+            log.error("Music not loaded");
+            return;
+        }
+        Music music = resourceService.getAsset(BACKGROUND_MUSIC, Music.class);
+        music.setLooping(true);
+        if (!UserSettings.get().mute) {
+            music.setVolume(UserSettings.get().musicVolume);
+        } else {
+            music.setVolume(0);
+        }
+        music.play();
+    }
+
+    protected void stopBackgroundMusic() {
+        ServiceLocator.getResourceService().getAsset(BACKGROUND_MUSIC, Music.class).stop();
+    }
+
+    // Asset-related methods moved from MainGameArea
+    @Override
+    protected String[] getSoundFilepaths() {
+        return new String[]{
+                "sounds/Impact4.ogg"
+        };
+    }
+
+    @Override
+    protected String[] getTextureAtlasFilepaths() {
+        return new String[]{
+                "images/terrain_iso_grass.atlas",
+                "skins/levels/level1/level1_skin.atlas",
+                "skins/levels/level2/level2_skin.atlas",
+                "skins/levels/level3/level3_skin.atlas",
+                "skins/minimap/minimap.atlas"
+        };
+    }
+
+    @Override
+    protected String[] getTextureFilepaths() {
+        List<String> filepaths = new ArrayList<>();
+        String[] commonTextures = {
+                "images/box_boy_leaf.png",
+                "images/rounded_door_v.png",
+                "images/rounded_door_h.png",
+                "images/staircase.png",
+                "skins/levels/level1/level1_skin.png",
+                "skins/levels/level2/level2_skin.png",
+                "skins/levels/level3/level3_skin.png",
+                "skins/minimap/minimap.png"
+        };
+        Collections.addAll(filepaths, commonTextures);
+
+        for (int level = 1; level <= 3; level++) {
+            filepaths.add("images/tile_1_level" + level + ".png");
+            filepaths.add("images/tile_2_level" + level + ".png");
+            filepaths.add("images/tile_3_level" + level + ".png");
+            filepaths.add("images/tile_4_level" + level + ".png");
+            filepaths.add("images/tile_5_level" + level + ".png");
+            filepaths.add("images/tile_6_level" + level + ".png");
+            filepaths.add("images/tile_7_level" + level + ".png");
+            filepaths.add("images/tile_8_level" + level + ".png");
+            filepaths.add("images/tile_middle_level" + level + ".png");
+            filepaths.add("images/tile_broken1_level" + level + ".png");
+            filepaths.add("images/tile_broken2_level" + level + ".png");
+            filepaths.add("images/tile_broken3_level" + level + ".png");
+            filepaths.add("images/tile_blood_level" + level + ".png");
+            filepaths.add("images/tile_staircase_level" + level + ".png");
+        }
+
+        // Convert the list to an array and return
+        return filepaths.toArray(new String[0]);
+    }
+    @Override
+    protected String[] getMusicFilepaths() {
+        return new String[]{BACKGROUND_MUSIC};
+    }
+
 }
